+++ src/services/schema-export/tests/models/all/cs/OptionalTypes.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using Realms;
+using MongoDB.Bson;
 
 namespace MyProject.Models
 {
@@ -30,6 +31,12 @@ namespace MyProject.Models
         [MapTo("dataOptional")]
         public byte[] DataOptional { get; set; }
 
+        [MapTo("objectIdOptional")]
+        public ObjectId? ObjectIdOptional { get; set; }
+
+        [MapTo("decimal128Optional")]
+        public Decimal128? Decimal128Optional { get; set; }
+
         [MapTo("objectOptional")]
         public RequiredTypes ObjectOptional { get; set; }
 
@@ -53,5 +60,11 @@ namespace MyProject.Models
 
         [MapTo("dataOptionalArray")]
         public IList<byte[]> DataOptionalArray { get; }
+
+        [MapTo("objectIdOptionalArray")]
+        public IList<ObjectId?> ObjectIdOptionalArray { get; }
+
+        [MapTo("decimal128OptionalArray")]
+        public IList<Decimal128?> Decimal128OptionalArray { get; }
     }
 }