+++ src/services/schema-export/tests/models/all/cs/RequiredTypes.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using Realms;
+using MongoDB.Bson;
 
 namespace MyProject.Models
 {
@@ -32,6 +33,12 @@ namespace MyProject.Models
         [MapTo("dataRequired")]
         public byte[] DataRequired { get; set; }
 
+        [MapTo("objectIdRequired")]
+        public ObjectId ObjectIdRequired { get; set; }
+
+        [MapTo("decimal128Required")]
+        public Decimal128 Decimal128Required { get; set; }
+
         [MapTo("boolRequiredArray")]
         public IList<bool> BoolRequiredArray { get; }
 
@@ -54,6 +61,12 @@ namespace MyProject.Models
         [Required]
         [MapTo("dataRequiredArray")]
         public IList<byte[]> DataRequiredArray { get; }
+
+        [MapTo("objectIdRequiredArray")]
+        public IList<ObjectId> ObjectIdRequiredArray { get; }
+
+        [MapTo("decimal128RequiredArray")]
+        public IList<Decimal128> Decimal128RequiredArray { get; }
 
         [MapTo("objectRequiredArray")]
         public IList<RequiredTypes> ObjectRequiredArray { get; }